исправления4

+++ commerce/shopping-cart/src/main/java/ru/yandex/practicum/controller/ShoppingCartController.java
@@ -2,16 +2,13 @@ package ru.yandex.practicum.controller;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
+import jakarta.validation.constraints.PositiveOrZero;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.dto.cart.ChangeProductQuantityRequest;
 import ru.yandex.practicum.dto.cart.ShoppingCartDto;
-import ru.yandex.practicum.exception.NotAuthorizedUserException;
 import ru.yandex.practicum.feign.ShoppingCartClient;
 import ru.yandex.practicum.service.CartService;
 
@@ -31,7 +28,6 @@ public class ShoppingCartController implements ShoppingCartClient {
     @Override
     @GetMapping
     public ShoppingCartDto getCart(@RequestParam @NotBlank String username) {
-        checkUser(username);
         return cartService.getCart(username);
     }
 
@@ -39,20 +35,18 @@ public class ShoppingCartController implements ShoppingCartClient {
     @PutMapping
     public ShoppingCartDto addProductToCart(
             @RequestParam @NotBlank String username,
-            @RequestBody
-            @Valid
-            @NotNull
-            @NotEmpty
-            Map<@NotNull UUID, @NotNull @Positive Long> products
+            @RequestBody(required = false)
+            Map<UUID, @PositiveOrZero Long> products
     ) {
-        checkUser(username);
+        if (products == null || products.isEmpty()) {
+            return cartService.getCart(username);
+        }
         return cartService.addProductToCart(username, products);
     }
 
     @Override
     @DeleteMapping
     public void deleteCart(@RequestParam @NotBlank String username) {
-        checkUser(username);
         cartService.deleteCart(username);
     }
 
@@ -60,13 +54,11 @@ public class ShoppingCartController implements ShoppingCartClient {
     @PostMapping("/remove")
     public ShoppingCartDto removeFromCart(
             @RequestParam @NotBlank String username,
-            @RequestBody
-            @Valid
-            @NotNull
-            @NotEmpty
-            Set<@NotNull UUID> productIds
+            @RequestBody(required = false) Set<UUID> productIds
     ) {
-        checkUser(username);
+        if (productIds == null || productIds.isEmpty()) {
+            return cartService.getCart(username);
+        }
         return cartService.removeFromCart(username, productIds);
     }
 
@@ -74,15 +66,11 @@ public class ShoppingCartController implements ShoppingCartClient {
     @PostMapping("/change-quantity")
     public ShoppingCartDto changeProductQuantity(
             @RequestParam @NotBlank String username,
-            @RequestBody @Valid @NotNull ChangeProductQuantityRequest request
+            @RequestBody(required = false) @Valid ChangeProductQuantityRequest request
     ) {
-        checkUser(username);
-        return cartService.changeProductQuantity(username, request);
-    }
-
-    private void checkUser(String username) {
-        if (username == null || username.isBlank()) {
-            throw new NotAuthorizedUserException("Пользователь не авторизован");
+        if (request == null) {
+            return cartService.getCart(username);
         }
+        return cartService.changeProductQuantity(username, request);
     }
 }
